Backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
SQLAlchemy==2.0.36
aiomysql==0.2.0
mysql-connector-python==9.0.0
python-dotenv==1.0.1
